+++ RTSPathingLib/Source/RTSPathingLib/Formation/Formation.h
@@ -31,10 +31,10 @@ namespace RTSPathingLib {
 
     size_t getUnitCategory() const;
     void   setUnitCategory(size_t);
-    float  getUnitDistributionWeight() const; // if multiple formation nodes have the same category, this can adjust the distribution weighting (higher numbers->more Units)
-    void   setUnitDistributionWeight(float v);
-    float  getUnitPadding() const; //how much the units are appart from eachother
-    void   setUnitPadding(float padding);
+    double  getUnitDistributionWeight() const; // if multiple formation nodes have the same category, this can adjust the distribution weighting (higher numbers->more Units)
+    void   setUnitDistributionWeight(double v);
+    double  getUnitPadding() const; //how much the units are appart from eachother
+    void   setUnitPadding(double padding);
 
     void setShape(std::unique_ptr<FormationShape> shape);
     FormationShape& getShape();
@@ -47,10 +47,10 @@ namespace RTSPathingLib {
     bool   overwriteWidthWithInterfaceWidth = false;
     bool   rotateWithInterface              = false;
     bool   linkSizeWithParent               = false;
-    float  unitPadding                      = 1.1f;
+    double  unitPadding                      = 1.1;
 
     size_t unitCategory = 0;
-    float  unitDistributionWeight = 1;
+    double  unitDistributionWeight = 1;
 
     std::vector<std::unique_ptr<Formation>> children;
     const Formation* parent = nullptr;
