+++ RTSPathingLib/Source/RTSPathingLib/Util/Geometry2D.h
@@ -6,6 +6,6 @@
 namespace RTSPathingLib {
   class Geometry2D {
   public:
-    static bool LineLineIntersect(glm::vec2 p1, glm::vec2 p2, glm::vec2 q1, glm::vec2 q2, glm::vec2& intersection);
+    static bool LineLineIntersect(glm::dvec2 p1, glm::dvec2 p2, glm::dvec2 q1, glm::dvec2 q2, glm::dvec2& intersection);
   };
 }
