+++ RTSPathingLib/Source/RTSPathingLib/Util/HexagonGrid/HexGridSvg.cpp
@@ -7,12 +7,12 @@
 #include "Util/svg.h"
 
 namespace RTSPathingLib {
-  std::vector<svg> HexGridSvg::write(const TopFlatHexGrid<bool>& grid, float scale) {
+  std::vector<svg> HexGridSvg::write(const TopFlatHexGrid<bool>& grid, double scale) {
     std::vector<svg> toWrite;
     std::vector<svg> overlay;
     for (size_t i = 0; i < grid.data.size(); i++) {
       svg v;
-      glm::vec2 hexCenter = HexGridUtils::getTopFlatOffset(glm::ivec2(i % grid.dimension.x, i / grid.dimension.x), glm::vec2(0, 0), scale);
+      glm::dvec2 hexCenter = HexGridUtils::getTopFlatOffset(glm::ivec2(i % grid.dimension.x, i / grid.dimension.x), glm::dvec2(0, 0), scale);
       v.streak = getHexagon(hexCenter, scale);
       v.filled = grid.data[i];
       toWrite.push_back(v);
@@ -24,14 +24,14 @@ namespace RTSPathingLib {
     return toWrite;
   }
 
-  std::vector<glm::vec2> HexGridSvg::getHexagon(const glm::vec2& offset, float radius) {
-    std::vector<glm::vec2> result;
+  std::vector<glm::dvec2> HexGridSvg::getHexagon(const glm::dvec2& offset, double radius) {
+    std::vector<glm::dvec2> result;
 
     for (int i = 0; i < 7; i++) {
-      glm::vec2 point;
-      float angle = (i) * 2.0f * glm::pi<float>() / 6.0f;
-      point.x = radius * glm::cos(angle) * 0.5f;
-      point.y = radius * glm::sin(angle) * 0.5f;
+      glm::dvec2 point;
+      double angle = (i) * 2.0 * glm::pi<double>() / 6.0;
+      point.x = radius * glm::cos(angle) * 0.5;
+      point.y = radius * glm::sin(angle) * 0.5;
       result.push_back(point + offset);
     }
     return result;
