+++ RTSPathingLib/Source/RTSPathingLib/Util/HexagonGrid/HexGridUtils.h
@@ -6,7 +6,7 @@
 namespace RTSPathingLib {
   class HexGridUtils {
   public:
-    static glm::vec2 getTopFlatOffset(const glm::ivec2& position, const glm::vec2& offset, float scale);
+    static glm::dvec2 getTopFlatOffset(const glm::ivec2& position, const glm::dvec2& offset, double scale);
     static std::vector<glm::ivec2> getSurroundedTiles(glm::ivec2 pos);
   };
 }
